Backend/requirements.txt
flask
werkzeug
google-genai
python-dotenv
Pillow
